updated organizations with new columns for social profiles. added a validation rule to the organizations' Revoke() method, users cannot be removed from OnRecord. updated the email template, removed the bubble image, replaced the OnRecord text with our logo and reduced cell margins. updated all copyright years to 2019. added account type filter controls to the organizations edit view. added an error handler to the remove button, in the organizations edit view (revoking users). hide the remove button for OnRecord users.

+++ WebApp/Controllers/OrganisationsController.cs
@@ -149,6 +149,12 @@ namespace WebApi.Controllers
             createOrganisation.LogoUrl = value.LogoUrl;
             createOrganisation.TermsAndConditions = value.TermsAndConditions;
             createOrganisation.RequiresAgreement = value.RequiresAgreement;
+            createOrganisation.FacebookId = value.FacebookId;
+            createOrganisation.TwitterId = value.TwitterId;
+            createOrganisation.InstagramId = value.InstagramId;
+            createOrganisation.SkypeId = value.SkypeId;
+            createOrganisation.LinkedinUrl = value.LinkedinUrl;
+            createOrganisation.YouTubeUrl = value.YouTubeUrl;
             createOrganisation.DefaultCalendarId = CalendarsRepository.Gregorian.Id;
             createOrganisation.DefaultLanguageId = LanguagesRepository.English.Id;
 
@@ -191,6 +197,12 @@ namespace WebApi.Controllers
             organisation.LogoUrl = value.LogoUrl;
             organisation.TermsAndConditions = value.TermsAndConditions;
             organisation.RequiresAgreement = value.RequiresAgreement;
+            organisation.FacebookId = value.FacebookId;
+            organisation.TwitterId = value.TwitterId;
+            organisation.InstagramId = value.InstagramId;
+            organisation.SkypeId = value.SkypeId;
+            organisation.LinkedinUrl = value.LinkedinUrl;
+            organisation.YouTubeUrl = value.YouTubeUrl;
 
             try
             {
@@ -292,6 +304,10 @@ namespace WebApi.Controllers
             // root users cannot be removed from an organization!
             if (orgUser.IsRootUser)
                 return BadRequest("Root users cannot be removed from organizations!");
+
+            // users cannot be removed from OnRecord
+            if (org.Name.Equals("OnRecord"))
+                return BadRequest("Users cannot be removed from OnRecord!");
 
             var subscriptionService = new SubscriptionService(UnitOfWork);
             subscriptionService.RemoveUserFromOrganization(org, orgUser);
